+++ src/Agent/NewRelic/Agent/Extensions/Providers/Wrapper/Owin/ResolveAppWrapper.cs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System;
+using System.Threading;
 using NewRelic.Agent.Api;
 using NewRelic.Agent.Extensions.Providers.Wrapper;
 using NewRelic.Reflection;
@@ -11,7 +12,7 @@ namespace NewRelic.Providers.Wrapper.Owin
     public class ResolveAppWrapper : IWrapper
     {
         public bool IsTransactionRequired => false;
-        private static bool _isNewRelicMiddlewareAdded;
+        private static int _isNewRelicMiddlewareAdded;
         private Func<object, object> _getBuilder;
         public Func<object, object> GetBuilder => _getBuilder ?? (_getBuilder = VisibilityBypasser.Instance.GeneratePropertyAccessor<object>("Microsoft.Owin.Hosting",
                 "Microsoft.Owin.Hosting.Engine.StartContext", "Builder"));
@@ -23,7 +24,7 @@ namespace NewRelic.Providers.Wrapper.Owin
 
         public AfterWrappedMethodDelegate BeforeWrappedMethod(InstrumentedMethodCall instrumentedMethodCall, IAgent agent, ITransaction transaction)
         {
-            if (!_isNewRelicMiddlewareAdded)
+            if (0 == Interlocked.CompareExchange(ref _isNewRelicMiddlewareAdded, 1, 0))
             {
                 var context = instrumentedMethodCall.MethodCall.MethodArguments[0];
 
@@ -36,7 +37,6 @@ namespace NewRelic.Providers.Wrapper.Owin
                 typeof(OwinStartupMiddleware), new object[] { agent }
                 });
 
-                _isNewRelicMiddlewareAdded = true;
             }
 
             return Delegates.NoOp;