+++ src/Agent/CHANGELOG.md
@@ -20,7 +20,7 @@ When the NetTCP Binding type is used in Windows Communication Foundation (WCF), 
 ### Fixes
 * Fixes an issue that may cause `InvalidCastException` due to an assembly version mismatch in Mvc3 instrumentation.
 * Fixes an async timing issue that can cause the end time of `Task`-returning methods to be determined incorrectly.
-* Fixes [#223](https://github.com/newrelic/newrelic-dotnet-agent/issues/223) that causes the agent incompatible with ASP.NET Core 5 RC1 applications.
+* Fixes [#223](https://github.com/newrelic/newrelic-dotnet-agent/issues/223) so the agent can be compatible with ASP.NET Core 5 RC1.
 
 ## [8.31] - 2020-08-17
 ### New Features